feat: add CartUpdateDto and update price handling in product models; modify middleware for improved session management

+++ Webshop.Data/Models/Product.cs
@@ -5,6 +5,6 @@
         public int Id { get; set; }
         public required string Name { get; set; }
         public required string Description { get; set; }
-        public required decimal Price { get; set; }
+        public required long Price { get; set; }
     }
 }
